Updates VM and builder with various fixes
Updates the VM to support compressed root images via xz. It adds logic to detect and decompress .xz images in the challenge directory, falling back to existing .raw or .img files, or 9p passthrough if no suitable image is found. It also fixes the qemu arguments construction.

Updates the builder to use specific versions of AFL++ and angr-management and disables Binary Ninja free installation in the Dockerfile. This ensures consistent and reproducible builds by pinning to known working versions.